refactor(twilio): split routing helpers
